feat : adding mapping model y exception

+++ src/main/java/com/fernandez/cars/service/CarServiceImpl.java
@@ -1,12 +1,16 @@
 package com.fernandez.cars.service;
 import com.fernandez.cars.dto.CarDTO;
+import com.fernandez.cars.exception.EntityNotFoundException;
 import com.fernandez.cars.model.Car;
 import com.fernandez.cars.repository.CarRepository;
+import com.fernandez.cars.utils.ObjectMapperUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -16,35 +20,35 @@ public class CarServiceImpl implements CarService {
     
     private final CarRepository carRepository;
 
-    private final  ModelMapper modelMapper;
-
     @Override
     public List<CarDTO> findAllCars() {
         log.info("CarServiceImpl[findAllCars]");
-        return carRepository.findAll()
-                .stream()
-                .map(car -> modelMapper.map(car,CarDTO.class))
-                .collect(Collectors.toList());
+        return ObjectMapperUtils.mapAll(carRepository.findAll(), CarDTO.class);
     }
 
     @Override
     public CarDTO findCarById(Long carId) {
-        log.info("CarServiceImpl[findCarById]");
-        Car carFounded = carRepository.getById(carId);
-        CarDTO carDTO = modelMapper.map(carFounded,CarDTO.class);
-        return carDTO;
+        log.info("CarServiceImpl[findCarById]", carId);
+        Optional<Car> car = carRepository.findById(carId);
+        if(!Optional.empty().isPresent()){
+            throw new EntityNotFoundException(Car.class, "id", carId.toString());
+        }
+        return ObjectMapperUtils.map(car,CarDTO.class);
     }
 
     @Override
     public CarDTO createOrUpdateCar(CarDTO carDTO) {
-        log.info("CarServiceImpl[createOrUpdateCar]");
-        Car carCreated = carRepository.save(modelMapper.map(carDTO,Car.class));
-        return modelMapper.map(carCreated,CarDTO.class);
+        log.info("CarServiceImpl[createOrUpdateCar]" , carDTO);
+        return ObjectMapperUtils.map(carRepository.save(ObjectMapperUtils.map(carDTO,Car.class)),CarDTO.class);
     }
 
     @Override
     public void deleteCarById(Long carId) {
-        log.info("CarServiceImpl[deleteCarById]");
+        log.info("CarServiceImpl[deleteCarById]", carId);
+        Optional<Car> car = carRepository.findById(carId);
+        if(!Optional.empty().isPresent()){
+            throw new EntityNotFoundException(Car.class, "id", carId.toString());
+        }
         carRepository.deleteById(carId);
     }
 }
